Improve custom route with flask blueprints
Also add crypto modules
