+++ backend/crypto/__init__.py
@@ -1,0 +1,2 @@
+from .token import generate_token
+from .hashing import default_backend as hasher
